--- project/app/src/main/java/com/example/cmput301w21t25/NonNegCountExperiment.java
@@ -1,4 +1,0 @@
-package com.example.cmput301w21t25;
-
-public class NonNegCountExperiment extends CountExperiment{
-}
