+++ project/app/src/main/java/com/example/cmput301w21t25/activities/ViewTrialActivity.java
@@ -1,4 +1,4 @@
-package com.example.cmput301w21t25;
+package com.example.cmput301w21t25.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
