+++ project/app/src/main/java/com/example/cmput301w21t25/adapters/CustomListExperiment.java
@@ -1,4 +1,4 @@
-package com.example.cmput301w21t25;
+package com.example.cmput301w21t25.adapters;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -9,6 +9,9 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
+import com.example.cmput301w21t25.R;
+import com.example.cmput301w21t25.experiments.Experiment;
 
 import java.util.ArrayList;
