+++ project/app/src/main/java/com/example/cmput301w21t25/experiments/CountExperiment.java
@@ -1,0 +1,6 @@
+package com.example.cmput301w21t25.experiments;
+
+import com.example.cmput301w21t25.experiments.Experiment;
+
+public class CountExperiment extends Experiment {
+}
