+++ project/app/src/main/java/com/example/cmput301w21t25/experiments/NonNegCountExperiment.java
@@ -1,0 +1,4 @@
+package com.example.cmput301w21t25.experiments;
+
+public class NonNegCountExperiment extends CountExperiment {
+}
