+++ project/app/src/main/java/com/example/cmput301w21t25/trials/BinomialTrial.java
@@ -1,4 +1,4 @@
-package com.example.cmput301w21t25;
+package com.example.cmput301w21t25.trials;
 
 /**
  * @author Samadhi
