+++ project/app/src/main/res/layout/activity_create_experiment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CreateExperimentActivity">
+    tools:context=".activities.CreateExperimentActivity">
 
     <TextView
         android:id="@+id/textViewNameQueryId"
