+++ project/app/src/main/res/layout/activity_temp_right.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TempRightActivity">
+    tools:context=".activities.TempRightActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
